add more detail for log
