+++ utils/resolvconf_test.go
@@ -56,24 +56,3 @@ func TestGetSearchDomains(t *testing.T) {
 		assert.ElementsMatch(t, domains, tc.expected)
 	}
 }
-
-func TestIsYaml(t *testing.T) {
-	testCases := []struct {
-		input    string
-		expected bool
-	}{
-		{
-			input:    "a.yaml",
-			expected: true,
-		}, {
-			input:    "a.yml",
-			expected: true,
-		}, {
-			input:    "a.txt",
-			expected: false,
-		},
-	}
-	for _, tc := range testCases {
-		assert.Equal(t, IsYaml(tc.input), tc.expected)
-	}
-}
