+++ utils/utils.go
@@ -15,6 +15,7 @@ package utils
 
 import (
 	"os"
+	"path/filepath"
 	"regexp"
 	"strings"
 	"time"
@@ -119,4 +120,13 @@ func SetEnvVariables(variables map[string]string) error {
 		}
 	}
 	return nil
+}
+
+// IsYaml checks whether the file is yaml or not.
+func IsYaml(fileName string) bool {
+	extension := strings.ToLower(filepath.Ext(fileName))
+	if extension == ".yaml" || extension == ".yml" {
+		return true
+	}
+	return false
 }
